Removed internal namespace from Networkgraph.
